add Vender/nova venda

+++ src/br/maua/models/Cadastro.java
@@ -6,11 +6,11 @@ import br.maua.enumerates.FormaPagamento;
 public class Cadastro {
     private String id;
     private String descricao;
-    private int valor;
+    private float valor;
     private FormaPagamento formaPagamento;
     private Estado estado;
 
-    public Cadastro(String id, String descricao, int valor, FormaPagamento formaPagamento, Estado estado) {
+    public Cadastro(String id, String descricao, float valor, FormaPagamento formaPagamento, Estado estado) {
         this.id = id;
         this.descricao = descricao;
         this.valor = valor;
